move real impls to cmd to avoid coverage
