--- pkg/entrypoint/postwriter.go
@@ -1,42 +1,0 @@
-/*
-Copyright 2019 The Knative Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package entrypoint
-
-import (
-	"log"
-	"os"
-)
-
-// PostWriter encapsulates writing a file when complete.
-type PostWriter interface {
-	// Write writes to the path when complete.
-	Write(file string)
-}
-
-// RealPostWriter actually writes files.
-type RealPostWriter struct{}
-
-var _ PostWriter = (*RealPostWriter)(nil)
-
-func (*RealPostWriter) Write(file string) {
-	if file == "" {
-		return
-	}
-	if _, err := os.Create(file); err != nil {
-		log.Fatalf("Creating %q: %v", file, err)
-	}
-}
